requirements.txt
numpy==2.0.2
Panda3D==1.10.14
pygame==2.6.1
PyOpenGL==3.1.7
